move constructor tests into separated files
