+++ packages/messaging-api-messenger/src/__tests__/MessengerClient.spec.js
@@ -10,12 +10,6 @@ const USER_ID = '1QAZ2WSX';
 const ACCESS_TOKEN = '1234567890';
 const APP_SECRET = '1WDVGY78';
 
-const createMock = () => {
-  const client = new MessengerClient(ACCESS_TOKEN);
-  const mock = new MockAdapter(client.axios);
-  return { client, mock };
-};
-
 let axios;
 let _create;
 beforeEach(() => {
@@ -27,78 +21,11 @@ afterEach(() => {
   axios.create = _create;
 });
 
-describe('connect', () => {
-  it('create axios with default graphAPI version', () => {
-    axios.create = jest.fn();
-    MessengerClient.connect(ACCESS_TOKEN);
-
-    expect(axios.create).toBeCalledWith({
-      baseURL: 'https://graph.facebook.com/v2.11/',
-      headers: { 'Content-Type': 'application/json' },
-    });
-  });
-
-  it('create axios with custom graphAPI version', () => {
-    axios.create = jest.fn();
-    MessengerClient.connect(ACCESS_TOKEN, '2.6');
-
-    expect(axios.create).toBeCalledWith({
-      baseURL: 'https://graph.facebook.com/v2.6/',
-      headers: { 'Content-Type': 'application/json' },
-    });
-  });
-});
-
-describe('constructor', () => {
-  it('create axios with default graphAPI version', () => {
-    axios.create = jest.fn();
-    new MessengerClient(ACCESS_TOKEN); // eslint-disable-line no-new
-
-    expect(axios.create).toBeCalledWith({
-      baseURL: 'https://graph.facebook.com/v2.11/',
-      headers: { 'Content-Type': 'application/json' },
-    });
-  });
-
-  it('create axios with custom graphAPI version', () => {
-    axios.create = jest.fn();
-    new MessengerClient(ACCESS_TOKEN, '2.6'); // eslint-disable-line no-new
-
-    expect(axios.create).toBeCalledWith({
-      baseURL: 'https://graph.facebook.com/v2.6/',
-      headers: { 'Content-Type': 'application/json' },
-    });
-  });
-});
-
-describe('#version', () => {
-  it('should return version of graph api', () => {
-    expect(new MessengerClient(ACCESS_TOKEN).version).toEqual('2.11');
-    expect(new MessengerClient(ACCESS_TOKEN, 'v2.6').version).toEqual('2.6');
-    expect(new MessengerClient(ACCESS_TOKEN, '2.6').version).toEqual('2.6');
-    expect(() => {
-      // eslint-disable-next-line no-new
-      new MessengerClient(ACCESS_TOKEN, 2.6);
-    }).toThrow('Type of `version` must be string.');
-  });
-});
-
-describe('#axios', () => {
-  it('should return underlying http client', () => {
-    const client = new MessengerClient(ACCESS_TOKEN);
-    expect(client.axios.get).toBeDefined();
-    expect(client.axios.post).toBeDefined();
-    expect(client.axios.put).toBeDefined();
-    expect(client.axios.delete).toBeDefined();
-  });
-});
-
-describe('#accessToken', () => {
-  it('should return underlying access token', () => {
-    const client = new MessengerClient(ACCESS_TOKEN);
-    expect(client.accessToken).toBe(ACCESS_TOKEN);
-  });
-});
+const createMock = () => {
+  const client = new MessengerClient(ACCESS_TOKEN);
+  const mock = new MockAdapter(client.axios);
+  return { client, mock };
+};
 
 describe('page info', () => {
   describe('#getPageInfo', () => {
