+++ packages/messaging-api-telegram/src/__tests__/TelegramClient-constructor.spec.js
@@ -1,0 +1,72 @@
+import TelegramClient from '../TelegramClient';
+
+const ACCESS_TOKEN = '123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11';
+
+describe('connect', () => {
+  let axios;
+  let _create;
+  beforeEach(() => {
+    axios = require('axios'); // eslint-disable-line global-require
+    _create = axios.create;
+  });
+
+  afterEach(() => {
+    axios.create = _create;
+  });
+
+  it('create axios with Telegram API', () => {
+    axios.create = jest.fn();
+    TelegramClient.connect(ACCESS_TOKEN);
+
+    expect(axios.create).toBeCalledWith({
+      baseURL:
+        'https://api.telegram.org/bot123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11/',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+  });
+});
+
+describe('constructor', () => {
+  let axios;
+  let _create;
+  beforeEach(() => {
+    axios = require('axios'); // eslint-disable-line global-require
+    _create = axios.create;
+  });
+
+  afterEach(() => {
+    axios.create = _create;
+  });
+
+  it('create axios with Telegram API', () => {
+    axios.create = jest.fn();
+    new TelegramClient(ACCESS_TOKEN); // eslint-disable-line no-new
+
+    expect(axios.create).toBeCalledWith({
+      baseURL:
+        'https://api.telegram.org/bot123456:ABC-DEF1234ghIkl-zyx57W2v1u123ew11/',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+  });
+});
+
+describe('#axios', () => {
+  it('should return underlying http client', () => {
+    const client = new TelegramClient(ACCESS_TOKEN);
+    expect(client.axios.get).toBeDefined();
+    expect(client.axios.post).toBeDefined();
+    expect(client.axios.put).toBeDefined();
+    expect(client.axios.delete).toBeDefined();
+  });
+});
+
+describe('#accessToken', () => {
+  it('should return underlying access token', () => {
+    const client = new TelegramClient(ACCESS_TOKEN);
+    expect(client.accessToken).toBe(ACCESS_TOKEN);
+  });
+});
