+++ packages/messaging-api-viber/src/__tests__/ViberClient.spec.js
@@ -17,75 +17,6 @@ const createMock = () => {
   return { client, mock };
 };
 
-describe('connect', () => {
-  let axios;
-  let _create;
-  beforeEach(() => {
-    axios = require('axios'); // eslint-disable-line global-require
-    _create = axios.create;
-  });
-
-  afterEach(() => {
-    axios.create = _create;
-  });
-
-  it('create axios with Viber API', () => {
-    axios.create = jest.fn();
-    ViberClient.connect(AUTH_TOKEN);
-
-    expect(axios.create).toBeCalledWith({
-      baseURL: 'https://chatapi.viber.com/pa/',
-      headers: {
-        'Content-Type': 'application/json',
-        'X-Viber-Auth-Token': AUTH_TOKEN,
-      },
-    });
-  });
-});
-
-describe('constructor', () => {
-  let axios;
-  let _create;
-  beforeEach(() => {
-    axios = require('axios'); // eslint-disable-line global-require
-    _create = axios.create;
-  });
-
-  afterEach(() => {
-    axios.create = _create;
-  });
-
-  it('create axios with Viber API', () => {
-    axios.create = jest.fn();
-    new ViberClient(AUTH_TOKEN); // eslint-disable-line no-new
-
-    expect(axios.create).toBeCalledWith({
-      baseURL: 'https://chatapi.viber.com/pa/',
-      headers: {
-        'Content-Type': 'application/json',
-        'X-Viber-Auth-Token': AUTH_TOKEN,
-      },
-    });
-  });
-});
-
-describe('#axios', () => {
-  it('should return underlying http client', () => {
-    const client = new ViberClient(AUTH_TOKEN);
-    expect(client.axios.get).toBeDefined();
-    expect(client.axios.post).toBeDefined();
-    expect(client.axios.put).toBeDefined();
-    expect(client.axios.delete).toBeDefined();
-  });
-});
-
-describe('#accessToken', () => {
-  it('should return underlying access token', () => {
-    const client = new ViberClient(AUTH_TOKEN);
-    expect(client.accessToken).toBe(AUTH_TOKEN);
-  });
-});
-
 describe('webhooks', () => {
   describe('#setWebhook', () => {
     it('should response event_types was set', async () => {
